+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
+Generiertes changelog template von gpt5 aus der benannten website:
+
 Alle Änderungen an diesem Projekt werden in dieser Datei dokumentiert.
 
 Das Format basiert auf [Keep a Changelog](https://keepachangelog.com/de/1.1.0/),
@@ -7,10 +9,12 @@ und dieses Projekt hält sich an [Semantic Versioning](https://semver.org/spec/v
 ## [Unreleased]
 
 ### Added
-- 
+- WebcamDeviceHandler stubs
+- Cookie stubs and definitions
+- Webcam hardware works now. Isolated the code by functionality for implementation in devicehandler
 
 ### Changed
-- 
+- Nothing. Removed all the relevant files from the build process so i dont get any errors.
 
 ### Deprecated
 -